requirements.txt
absl-py==1.4.0
addict==2.4.0
altair==4.2.2
attrs==23.1.0
backports.zoneinfo==0.2.1
basicsr==1.3.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cycler==0.11.0
cython==0.29.33
decorator==5.1.1
einops==0.6.1
entrypoints==0.4
filelock==3.12.0
fonttools==4.38.0
future==0.18.3
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.1
google-auth-oauthlib==0.4.6
grpcio==1.54.2
h5py==3.8.0
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lmdb==1.4.1
lpips==0.1.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.3
mdurl==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
natsort==8.3.1
networkx==2.6.3
numpy
oauthlib==3.2.2
opencv-python-headless
packaging==23.1
pandas
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
streamlit==1.22.0
tb-nightly==2.12.0a20230113
tenacity==8.2.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2021.11.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.8.1
torchvision==0.9.1
tornado==6.2
tqdm==4.65.0
typing_extensions==4.3.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.2.3
wget
wincertstore==0.2
yacs==0.1.8
yapf==0.33.0
zipp==3.15.0